Lift the entire web server machinery!
So that we can do "ask" in the middle of our HTML rendering to retrieve
some context. Some might say, after that fact, that this would have been
better achieved by passing in a context argument. However, this is probably
neater in the long run, as it potentially simplifies refactoring a bit.

Only time will tell!

The proof of this whole pudding is in a nice little "foo" we get in the top left
of each page!