Check session active before showing page
We (eventually) want to let you log off other sessions that may have been
compromised (e.g. if you left one logged in on a shared machine).
Therefore, we need to check each time you make a request that the session
UUID is still in the list of active sessions. This means that when we
redirect you when you first sign in, we can't just assume the view has
updated yet to reflect your sign in. That is, we need to wait on you
becoming active.

This commit changes around a bunch of stuff, some of which is just WIP:
 - We return the latest version of a stream when we log events (this isn't
   useful for waiting on sessions, but will be useful when waiting for
   account updates where no particular lifecycle change is observed)
 - We start trying to generalise the view interface a bit more using
   IsView, so that we can build different objects that provide different
   functionality (like waiting for particular events)
 - We stop trying to pass all our views to a single runner (because getting
   a heterogeneous list of read views is a pain that's not worth bothering
   with.
 - We narrow the scope of DB access in read views by making the type that
   client code sees just IO. I had some connection contention where stuff
   blocked whilst someone else was waiting. Not sure about this change!
 - We stop handing back a cookie from the sign in command handler - instead
   we just report status. Something else then goes and waits for the cookie
   from the process manager. Not sure how good this pattern is, but it
   stops you just returning it to the user without waiting on the view.
 - We provide a a waiting function that waits until a session UUID is in
   the active sessions table and then generates you an appropriate cookie.

Phew, bonus dedication points if you read through all of this!